requirements.txt
django
djangorestframework
psycopg2-binary
pillow
python-slugify
python-decouple
